+++ packages/packages/core/editor-editing-panel/src/mcp/classes/init-element-classes-mcp.ts
@@ -2,7 +2,7 @@ import { getMCPByDomain } from '@elementor/editor-mcp';
 import { stylesRepository } from '@elementor/editor-styles-repository';
 import { z } from '@elementor/schema';
 
-import { doApplyClasses, doGetAppliedClasses, doUnapplyClass } from '../../components/css-classes/apply-unapply-utils';
+import { doApplyClasses, doGetAppliedClasses, doUnapplyClass } from '../../apply-unapply-actions';
 
 const TOOL_NAME = 'element_classes';
